refactor(core): fix layering inversions between core, config, and adapters
- invert core/layout dependency on integrations/sessionstate by moving the
  pure snapshot data model to internal/core/sessionstate; the integrations
  package re-exports it via type aliases so existing importers are unchanged
- make config a leaf by relocating the provider registry to
  internal/aiprovider; internal/core/aiprovider stays as a re-export facade
  so app/adapter import paths keep working
- move theme-dependent usage HUD bar rendering out of core/usage into
  internal/ui/render (tests move with it)
