+++ internal/aiprovider/registry_test.go
