+++ internal/core/layout/layout.go
@@ -12,7 +12,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/crevissepartners/projmux/internal/integrations/sessionstate"
+	"github.com/crevissepartners/projmux/internal/core/sessionstate"
 )
 
 const (
