+++ internal/ui/render/usagebar_test.go
@@ -1,4 +1,4 @@
-package usage
+package render
 
 import (
 	"strings"
